feature: change cluster manager return code to digital, add region to quota info; issue #904

+++ bcs-services/bcs-cluster-manager/internal/actions/cluster/get.go
@@ -67,7 +67,7 @@ func (ga *GetAction) getCluster() error {
 	return nil
 }
 
-func (ga *GetAction) setResp(code uint64, msg string) {
+func (ga *GetAction) setResp(code uint32, msg string) {
 	ga.resp.Code = code
 	ga.resp.Message = msg
 	ga.resp.Result = (code == types.BcsErrClusterManagerSuccess)
