+++ bcs-services/bcs-cluster-manager/internal/actions/clustercredential/delete.go
@@ -48,7 +48,7 @@ func (da *DeleteAction) deleteCredential() error {
 	return da.model.DeleteClusterCredential(da.ctx, da.req.ServerKey)
 }
 
-func (da *DeleteAction) setResp(code uint64, msg string) {
+func (da *DeleteAction) setResp(code uint32, msg string) {
 	da.resp.Code = code
 	da.resp.Message = msg
 	da.resp.Result = (code == types.BcsErrClusterManagerSuccess)
